+++ src/patch/core/GHideRaceUI.zig
@@ -4,6 +4,7 @@ const app = @import("../appinfo.zig");
 const GlobalSt = app.GLOBAL_STATE;
 const GlobalFn = app.GLOBAL_FUNCTION;
 const GLOBAL_STATE = &@import("Global.zig").GLOBAL_STATE;
+const workingOwner = @import("Hook.zig").PluginState.workingOwner;
 
 const rq = @import("racer").Quad;
 const rg = @import("racer").Global;
@@ -13,17 +14,15 @@ const ActiveState = @import("../util/active_state.zig").ActiveState;
 
 // FIXME: resolve clashing with practice mode indicators (should not hide them
 // even when everything else is). also makes lighting effects disappear
-// FIXME: convert to internal owner IDs once merged with custom trigger branch
 // TODO: check with LP if there is a different way of hiding race UI
 
 pub const HideRaceUI = extern struct {
     var hidden: bool = false;
-    var owner: ?[*:0]const u8 = null;
+    var owner: ?u16 = null;
 
     var paused: ActiveState = .On; // force .JustOff on first frame
 
-    /// @return request processed successfully
-    pub fn hide(o: [*:0]const u8) bool {
+    pub fn hide(o: u16) bool {
         if (hidden or owner != null) return false;
         writeHide(true);
         owner = o;
@@ -31,10 +30,8 @@ pub const HideRaceUI = extern struct {
         return true;
     }
 
-    /// @return request processed successfully
-    pub fn unhide(o: [*:0]const u8) bool {
-        const o_len = std.mem.len(o);
-        if (!hidden or !std.mem.eql(u8, owner.?[0..o_len], o[0..o_len])) return false;
+    pub fn unhide(o: u16) bool {
+        if (!hidden or owner != o) return false;
         writeHide(false);
         owner = null;
         hidden = false;
@@ -48,12 +45,24 @@ pub const HideRaceUI = extern struct {
         _ = mem.write(0x463580, u8, instruction); // top of Jdge0x20
         rq.QUAD_SKIP_RENDERING.* = @intFromBool(disable);
     }
-
-    /// @return game currently hiding race ui via api
-    pub fn isHidden() bool {
-        return hidden or owner != null;
-    }
 };
+
+// GLOBAL EXPORTS
+
+/// @return request processed successfully
+pub fn GHideRaceUIOn() bool {
+    return HideRaceUI.hide(workingOwner());
+}
+
+/// @return request processed successfully
+pub fn GHideRaceUIOff() bool {
+    return HideRaceUI.unhide(workingOwner());
+}
+
+/// @return game currently hiding race ui via api
+pub fn GHideRaceUIIsOn() bool {
+    return HideRaceUI.hidden or HideRaceUI.owner != null;
+}
 
 // HOOKS
 
@@ -61,7 +70,10 @@ pub fn OnInit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
 
 pub fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
 
-pub fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+pub fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+    if (HideRaceUI.owner) |o|
+        _ = HideRaceUI.unhide(o);
+}
 
 pub fn EarlyEngineUpdateB(gs: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     HideRaceUI.paused.update(rg.PAUSE_STATE.* > 0);
@@ -75,4 +87,9 @@ pub fn EarlyEngineUpdateB(gs: *GlobalSt, _: *GlobalFn) callconv(.C) void {
         .JustOff => HideRaceUI.writeHide(false),
         else => {},
     }
+}
+
+pub fn OnPluginDeinit(owner: u16) callconv(.C) void {
+    if (HideRaceUI.owner == owner)
+        _ = HideRaceUI.unhide(owner);
 }